Multithreaded implemented
Changed the socket to TCP, had problems with UDP and big messages that had to be sent in smaller packages (create a buffer). #4
Implemented a multithreaded tcp server, already tested with multiple clients using telnet.
This module already accepts at least one connection per time, parse the protocol received, execute the command and send it back to the backend.
It's also multithreaded.
So, I guess this closes #4