Backend works with TCP now
Some changes had to be made so the backend would communicate with tcp instead of udp. Some big changes inside sendAll function, there isn't any open/close of socket there anymore, only sending messages.
TCP connection starts when adding the host, it also binds backend with the host daemon. The host can be removed and it's connection will be closed. It's recommended to use method *closeAllConnection* at the end of the use of the program. These are all the updates for #3, there might be some small changes necessary to work with the #2.